app/requirements.txt
langchain-cli
langsmith
uvicorn
openai
langchainhub
pandas
langchain
python-dotenv
langchain-core
langchain-experimental
langchain-ollama
langchain-openai
fastapi
faiss-cpu
langchain-community
pymongo
motor
pydantic
pinecone



